Fixed console.log, started work on variables
I removed all the logDup stuff and followed your pattern of code injection. However I was not able to complete the overhaul of the variables tester because I'm worried about the current Frames stuff, in particular while I was running it it seemed to be showing multiple frames for functions I had only run once. Let's try to talk soon.